requirements.txt
blinker==1.8.2
boto3==1.35.46
botocore==1.35.46
click==8.1.7
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==3.0.2
python-dateutil==2.9.0.post0
s3transfer==0.10.3
six==1.16.0
urllib3==2.2.3
Werkzeug==3.0.4
python-dotenv
gunicorn